requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask-Migrate==4.0.5 